Fixes for the structure missalignments
Just a few quick fixes for your awesome decompiler. I've noticed the field offsets on the metadata files didn't match with il2cpp dumper's output and looking further into it with Ghidra and the other files revealed a few issues.
Looks like types with generic parameters are not located so I made a hack around it since I'm not familiar with the code (you should definitely change that) and the structs are not pieced together as they should, as I explained in the comment.
Very cool project, keep at it!