Core: DummyDll: Fix API consumers getting bad value types...
If types were defined out-of-order in the metadata, and anyone tried to use the generated assembly refs without saving the assemblies to disk first, some value types could be referenced not as value types, which was incorrect. Ensure we configure the hierarchy for VTs and Enums early to prevent this.